Reorganize runtimes
